refactor: abstracted addform logic into FormComponent
